Initial commit and working version.
This is basically a stripped down Fedora kickstart parser combined with blivet.
At the moment, this involves a modified version of pyanaconda/kickstart.py, at
least until most of the RedHat/Fedora-specifics are consolidated upstream.

I chose the license to be GPLv2+, because that's the license of Anaconda as
well, and I don't want to introduce licensing incompatibilities before all code
is actually rewritten from scratch.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>