Mount filesystems after successful partitioning.
We cannot use blivet.turnOnFilesystems() here, because for the NixOS installer,
we don't need cruft such as bindmounts and/or /dev, /proc, /sys and the like.
Thus, we traverse the device list by ourselves and mount only the "real"
filesystems that are actually tied to a mountpoint and not a special device.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>