Add option to skip partitioning and to mount only.
This sounds easier than it was, because blivet _usually_ has more information
about the state of the device tree. On NixOS, we don't want to keep any
additional state around and work exclusively on the information we got from the
kickstart file.

As you might have guessed, this requires guerrilla patching again:

One is for the blivet.devicelibs.mdraid.mdactivate() function, which seems to be
somewhat half-implemented in blivet. The original function has a keyword
argument called super_minor, which isn't implemented yet, but required if no
UUID is passed. We just silently skip the super_minor check, as mdadm is able to
bring up the device anyway.

Another one is not exactly monkey patching. Before calling setup() on the device
format, we need to traverse the device tree and force the "exists" attribute and
the device path to be enabled/set. If we don't do this, blivet will bail out
because the device paths/exists flags are set by the partitioning methods.
Unfortunately, and for obvious reasons, we can't run these in mount_only mode.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>