Improve setup.py and restructure directory layout.
We're preparing for nixpart 1.0, and we'll start by creating a better
directory layout, especially because we're going to implement unit tests
from the beginning.

Also the fields provided in setup.py were a bit terse and are a bit more
comprehensional now.

Note that the license still is GPL and we're going to change to MIT
license as soon as we got rid of all the code from Anaconda.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>