nix: Safely pass path to nix-instantiate
The previous version used str.format() to pass the file path to
nix-instantiate, which is a bit fragile for example if it contains
single or double quotes or even if it's relative.

We now use the --arg flag of nix-instantiate to pass the file in a way
that we don't need any escaping at all (the file path is directly passed
to execve() so neither shell nor Nix parsing rules apply.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>