nixpart/nix: Remove unused compiled regex
This was for a pretty stupid workaround I did to let it work with our
NixOS tests. The copyFileFromHost() sub is not safe against single
quotes (it did an "echo '$text' > $targetfile" without escaping $text).

So no idea why I did introduce this stupidity, but the XML text
declaration suddenly ended up with stripped single quotes like this:

<?xml version=1.0 encoding=utf-8?>

And this in turn caused ElementTree.fromstring() to fail parsing the XML
input.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>