storage: Handle the new size NixOS type
Size options are now no longer simply strings but either an integer
which specifies the size in bytes, just the string "fill" or an
attribute set of various size units and their sizes.

So something like { mib = 400; gib = 10; } is now essentially
Size("400 MiB") + Size("10 GiB").

The value "fill" is now the same as we had with "grow" before.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>