Switch to system.build.nixpart-spec and JSON
Instead of our complicated XML parsing and the instantiaton of Nix
expressions, we now have a much cleaner interface which is easier to
evaluate and build and also is less error-prone and simplifies our code.

Fortunately we already have the same Python structure than we had before
with our xml2python() function, so we don't need to change anything in
the actual build up of the device tree.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>