Add a test case for storage NixOS config
This is certainly one of the most expensive tests to begin with and is
there mainly to make sure nix-build is correctly invoked and whether we
get out valid JSON from it.

All further tests that will be implemented are going to be based on JSON
fixtures, so we don't need to have Nix or the Nix store available during
testing.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>