tests: Add a small size test for devtree module
The test is everything but comprehensive but it serves a start for all
the ridiculous mocking we need to do in order to prevent blivet from
actually inspecting real devices.

We cannot properly test "fill" for the size specification, because we'd
need to run DeviceTree.realize() beforehand. This would mean even more
mocking and that's really not worth it because we already have NixOS VM
tests for that.

So in the end, it's more about whether the input Nix expression (or
better: its JSON variant) is correctly converted into a blivet device
tree.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>