args: Add a -m/--mount GNU getopt style argument
This is a bit more involved because argparse can't handle such
arguments, see http://stackoverflow.com/q/40989413.

In our case, we didn't need to monkey patch anything because we have a
fairly simple set of options that don't need optional arguments.

We can get away by preprocessing the arguments before passing them to
the parse_args() method of argparse.ArgumentParser.

Unfortunately when it comes to formatting the mount option, we need to
override internal methods of the default HelpFormatter.

The main reason I went with a GNU style option is that we'd otherwise
need to split it into two separate arguments, like --mount and
--mount-point. Being used to things like --mount=/foo or -m/foo I find
it more natural to only use only one argument for that.

I also added a few tests, which now means, that the 'args' test module
now isn't a stub anymore.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>