Add rudimentary mounting support using labels
This still needs to undergo a lot of refactoring, because the way we
build up the device tree is top-down, but for mounting we need to go
both ways.

So for example if a matcher for a particular disk doesn't apply anymore,
we want to walk button-up to figure out the disk without the matcher and
also without the label - of course only if it's distinguishable.

Hence, we need to break apart the whole populate() method, especially so
that values coming from expr are clearly set apart and not part of a
blivet/Nix mishmash.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>