devtree: Switch to using a dict for devspecs
The device specifications still are in the "type.name" format for the
users of storage.* options on the NixOS side but now get converted into
an internal representation of {'name': ..., 'type': ...} tuples, so we
no longer need string concatenation and splitting in order to extract a
type or a name of a devspec.

However in order to be hashable in our internal storagetree dict, we
need to convert the devspec dict into a tuple (devspec2tuple). Nix
doesn't have tuples, only lists and it would be a bit awkward to convert
it into something like [type name].

In the end however we're going to refactor the populate() method anyway,
so we might have a dedicated DevSpec class for that.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>