devtree: Rename new_uuid to just uuid
This has been requested by @vpodzime in rhinstaller/blivet#537 and it
obviously didn't make sense to introduce a new keyword argument.

IIRC the reason for introducing it in the first place was that it might
conflict with the device tree populator. But when creating a new file
system it doesn't matter because we don't go through the populator
anyway and we also don't run _create on a pre-existing filesystem unless
we really want to format it, but then it's fine to set the UUID.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>