Fix generating invalid Nix attribute sets
If the data given to nixify() is a dictionary and it's empty, the result
is that we get *two* empty Nix attribute sets, like this:

fileSystems = {}{};

So either we'd need to return after appending {} to the output or just
remove the check for whether data is an empty dict.

I choose the latter as it decreases the amount of branches.

Thanks to @nh2 for reporting.

Signed-off-by: aszlig <aszlig@nix.build>
Fixes: #11